revert last erratic commit
