+++ IscDbc/TypesResultSet.cpp
@@ -34,28 +34,22 @@
 #include "IscStatement.h"
 #include "Value.h"
 
-using namespace Firebird;
-
 namespace IscDbcLibrary {
-/*
+
 #define SET_SQLVAR(index, name, type, prec, offset)			\
 {															\
-	XSQLVAR *var = ((XSQLDA*)*sqlda)->sqlvar + index - 1;	\
-	char *src = var->sqlname;							\
-	const char *dst = name;									\
-	do														\
-		*src++=*dst;										\
-	while(*dst++);											\
+	auto *var = sqlda->orgVar( index );						\
 															\
-	var->sqlname_length = src - var->sqlname - 1;			\
-	var->sqltype = type + 1;								\
+	var->sqlname = name;									\
+	var->sqltype = type;									\
+	var->isNullable = true;									\
 	var->sqllen = prec;										\
 	var->sqldata = (char*)offset;							\
 }															\
 
 #define SET_INDICATOR_VAL(col,type,isNull)  if ( isNull && (*(type*)(var[col].sqldata + sqldataOffsetPtr)) == -1 ) *var[col].sqlind = -1; else *var[col].sqlind = 0;
 #define SET_INDICATOR_STR(col)  if ( ((char*)(var[col].sqldata + sqldataOffsetPtr)) == NULL || !strlen(((char*)(var[col].sqldata + sqldataOffsetPtr))) ) *var[col].sqlind = -1; else *var[col].sqlind = (short)strlen((char*)(var[col].sqldata + sqldataOffsetPtr)) + 1;
-*/
+
 struct Types {
 	char	label;
 	short	lenTypeName;
@@ -84,59 +78,55 @@ struct Types {
 	short	typeIntervalPrecision;
     };
 
-#define F_NO_NULLS					0		// SQL_NO_NULLS
-#define F_NULLABLE					1		// SQL_NULLABLE
-#define F_CASE_SENSITIVE			1		// SQL_TRUE
-#define F_CASE_INSENSITIVE			0		// SQL_FALSE
-#define F_IS_SIGNED					1		// SQL_TRUE
-#define F_NOT_SIGNED				0		// SQL_FALSE
-#define F_NOT_NUMERIC				-1
-#define F_SEARCHABLE_EXCEPT_LIKE	2		// SQL_ALL_EXCEPT_LIKE
-#define F_SEARCHABLE				3		// SQL_SEARCHABLE
-#define F_UNSEARCHABLE				0		// SQL_UNSEARCHABLE
-#define F_MONEY						1		// SQL_TRUE
-#define F_NOT_MONEY					0		// SQL_FALSE
-#define F_NOT_AUTO_INCR				0		// SQL_FALSE
-#define F_UNSCALED					-1
+#define NO_NULLS			0		// SQL_NO_NULLS
+#define NULLABLE			1		// SQL_NULLABLE
+#define CASE_SENSITIVE		1		// SQL_TRUE
+#define CASE_INSENSITIVE	0		// SQL_FALSE
+#define IS_SIGNED			1		// SQL_TRUE
+#define NOT_SIGNED			0		// SQL_FALSE
+#define NOT_NUMERIC			-1
+#define SEARCHABLE_EXCEPT_LIKE  2	// SQL_ALL_EXCEPT_LIKE
+#define SEARCHABLE			3		// SQL_SEARCHABLE
+#define UNSEARCHABLE		0		// SQL_UNSEARCHABLE
+#define MONEY				1		// SQL_TRUE
+#define NOT_MONEY			0		// SQL_FALSE
+#define NOT_AUTO_INCR		0		// SQL_FALSE
+#define UNSCALED			-1
 
 #define TYPE_SQL_DATETIME	9
 
-#define ALPHA(type,code,prec) 0,sizeof(type)-1,type,code,prec,1,"'",1,"'",6,"length",F_NULLABLE,F_CASE_SENSITIVE,F_SEARCHABLE,F_NOT_NUMERIC,F_NOT_MONEY,F_NOT_NUMERIC,sizeof(type)-1,type,F_UNSCALED,F_UNSCALED,code,F_NOT_NUMERIC,F_NOT_NUMERIC,F_NOT_NUMERIC
-#define BLOB(type,code,prec,prefix,suffix,casesensitive) 0,sizeof(type)-1,type,code,prec,sizeof(prefix)-1,prefix,sizeof(suffix)-1,suffix,0,"",F_NULLABLE,casesensitive,F_UNSEARCHABLE,F_NOT_NUMERIC,F_NOT_MONEY,F_NOT_NUMERIC,sizeof(type)-1,type,F_UNSCALED,F_UNSCALED,code,F_NOT_NUMERIC,F_NOT_NUMERIC,F_NOT_NUMERIC
-#define NUMERIC_TINYINT(type,code,prec,attr,min,max,numprecradix) 0,sizeof(type)-1,type,code,prec,0,"",0,"",sizeof(attr)-1,attr,F_NULLABLE,F_CASE_INSENSITIVE,F_SEARCHABLE_EXCEPT_LIKE,F_IS_SIGNED,F_NOT_MONEY,F_NOT_AUTO_INCR,25,"CHAR CHARACTER SET OCTETS",min,max,code,F_NOT_NUMERIC,numprecradix,F_NOT_NUMERIC
-#define NUMERIC(type,code,prec,attr,min,max,numprecradix) 0,sizeof(type)-1,type,code,prec,0,"",0,"",sizeof(attr)-1,attr,F_NULLABLE,F_CASE_INSENSITIVE,F_SEARCHABLE_EXCEPT_LIKE,F_NOT_SIGNED,F_NOT_MONEY,F_NOT_AUTO_INCR,sizeof(type)-1,type,min,max,code,F_NOT_NUMERIC,numprecradix,F_NOT_NUMERIC
-#define DATE(type,code,prec,prefix,suffix,datetimesub) 0,sizeof(type)-1,type,code,prec,sizeof(prefix)-1,prefix,sizeof(suffix)-1,suffix,0,"",F_NULLABLE,F_CASE_INSENSITIVE,F_SEARCHABLE_EXCEPT_LIKE,F_NOT_NUMERIC,F_NOT_MONEY,F_NOT_NUMERIC,sizeof(type)-1,type,F_UNSCALED,F_UNSCALED,TYPE_SQL_DATETIME,datetimesub,F_NOT_NUMERIC,F_NOT_NUMERIC
-#define DATETIME(type,code,prec,prefix,suffix,datetimesub) 0,sizeof(type)-1,type,code,prec,sizeof(prefix)-1,prefix,sizeof(suffix)-1,suffix,0,"",F_NULLABLE,F_CASE_INSENSITIVE,F_SEARCHABLE_EXCEPT_LIKE,F_NOT_NUMERIC,F_NOT_MONEY,F_NOT_NUMERIC,sizeof(type)-1,type,0,4,TYPE_SQL_DATETIME,datetimesub,F_NOT_NUMERIC,F_NOT_NUMERIC
+#define ALPHA(type,code,prec) 0,sizeof(type)-1,type,code,prec,1,"'",1,"'",6,"length",NULLABLE,CASE_SENSITIVE,SEARCHABLE,NOT_NUMERIC,NOT_MONEY,NOT_NUMERIC,sizeof(type)-1,type,UNSCALED,UNSCALED,code,NOT_NUMERIC,NOT_NUMERIC,NOT_NUMERIC
+#define BLOB(type,code,prec,prefix,suffix,casesensitive) 0,sizeof(type)-1,type,code,prec,sizeof(prefix)-1,prefix,sizeof(suffix)-1,suffix,0,"",NULLABLE,casesensitive,UNSEARCHABLE,NOT_NUMERIC,NOT_MONEY,NOT_NUMERIC,sizeof(type)-1,type,UNSCALED,UNSCALED,code,NOT_NUMERIC,NOT_NUMERIC,NOT_NUMERIC
+#define NUMERIC_TINYINT(type,code,prec,attr,min,max,numprecradix) 0,sizeof(type)-1,type,code,prec,0,"",0,"",sizeof(attr)-1,attr,NULLABLE,CASE_INSENSITIVE,SEARCHABLE_EXCEPT_LIKE,IS_SIGNED,NOT_MONEY,NOT_AUTO_INCR,25,"CHAR CHARACTER SET OCTETS",min,max,code,NOT_NUMERIC,numprecradix,NOT_NUMERIC
+#define NUMERIC(type,code,prec,attr,min,max,numprecradix) 0,sizeof(type)-1,type,code,prec,0,"",0,"",sizeof(attr)-1,attr,NULLABLE,CASE_INSENSITIVE,SEARCHABLE_EXCEPT_LIKE,NOT_SIGNED,NOT_MONEY,NOT_AUTO_INCR,sizeof(type)-1,type,min,max,code,NOT_NUMERIC,numprecradix,NOT_NUMERIC
+#define DATE(type,code,prec,prefix,suffix,datetimesub) 0,sizeof(type)-1,type,code,prec,sizeof(prefix)-1,prefix,sizeof(suffix)-1,suffix,0,"",NULLABLE,CASE_INSENSITIVE,SEARCHABLE_EXCEPT_LIKE,NOT_NUMERIC,NOT_MONEY,NOT_NUMERIC,sizeof(type)-1,type,UNSCALED,UNSCALED,TYPE_SQL_DATETIME,datetimesub,NOT_NUMERIC,NOT_NUMERIC
+#define DATETIME(type,code,prec,prefix,suffix,datetimesub) 0,sizeof(type)-1,type,code,prec,sizeof(prefix)-1,prefix,sizeof(suffix)-1,suffix,0,"",NULLABLE,CASE_INSENSITIVE,SEARCHABLE_EXCEPT_LIKE,NOT_NUMERIC,NOT_MONEY,NOT_NUMERIC,sizeof(type)-1,type,0,4,TYPE_SQL_DATETIME,datetimesub,NOT_NUMERIC,NOT_NUMERIC
 
 static Types types [] = 
 {
 	ALPHA ("CHAR", JDBC_CHAR, MAX_CHAR_LENGTH),
 	ALPHA ("VARCHAR", JDBC_VARCHAR, MAX_VARCHAR_LENGTH),
-	BLOB ("BLOB SUB_TYPE TEXT", JDBC_LONGVARCHAR, MAX_BLOB_LENGTH, "'","'",F_CASE_SENSITIVE),
+	BLOB ("BLOB SUB_TYPE TEXT", JDBC_LONGVARCHAR, MAX_BLOB_LENGTH, "'","'",CASE_SENSITIVE),
 	ALPHA ("CHAR(x) CHARACTER SET UNICODE_FSS", JDBC_WCHAR, MAX_WCHAR_LENGTH),
 	ALPHA ("VARCHAR(x) CHARACTER SET UNICODE_FSS", JDBC_WVARCHAR, MAX_WVARCHAR_LENGTH),
-	BLOB ("BLOB SUB_TYPE TEXT CHARACTER SET UNICODE_FSS", JDBC_WLONGVARCHAR, MAX_WLONGVARCHAR_LENGTH, "'","'",F_CASE_SENSITIVE),
-	BLOB ("BLOB SUB_TYPE 0", JDBC_LONGVARBINARY, MAX_BLOB_LENGTH, "","",F_CASE_INSENSITIVE),
-	BLOB ("BLOB SUB_TYPE 0", JDBC_VARBINARY, MAX_BLOB_LENGTH, "","",F_CASE_INSENSITIVE),
-	BLOB ("BLOB SUB_TYPE 0", JDBC_BINARY, MAX_BLOB_LENGTH, "","",F_CASE_INSENSITIVE),
+	BLOB ("BLOB SUB_TYPE TEXT CHARACTER SET UNICODE_FSS", JDBC_WLONGVARCHAR, MAX_WLONGVARCHAR_LENGTH, "'","'",CASE_SENSITIVE),
+	BLOB ("BLOB SUB_TYPE 0", JDBC_LONGVARBINARY, MAX_BLOB_LENGTH, "","",CASE_INSENSITIVE),
+	BLOB ("BLOB SUB_TYPE 0", JDBC_VARBINARY, MAX_BLOB_LENGTH, "","",CASE_INSENSITIVE),
+	BLOB ("BLOB SUB_TYPE 0", JDBC_BINARY, MAX_BLOB_LENGTH, "","",CASE_INSENSITIVE),
 	NUMERIC ("NUMERIC", JDBC_NUMERIC, MAX_NUMERIC_LENGTH, "precision,scale", 0, MAX_NUMERIC_LENGTH, 10),
 	NUMERIC ("DECIMAL", JDBC_DECIMAL, MAX_DECIMAL_LENGTH, "precision,scale", 0, MAX_DECIMAL_LENGTH, 10),
 	NUMERIC ("INTEGER", JDBC_INTEGER, MAX_INT_LENGTH, "", 0, 0, 10),
 	NUMERIC ("SMALLINT", JDBC_TINYINT, MAX_SMALLINT_LENGTH, "", 0, 0, 10),
 	NUMERIC ("SMALLINT", JDBC_SMALLINT, MAX_SMALLINT_LENGTH, "", 0, 0, 10),	
-	NUMERIC ("DOUBLE PRECISION", JDBC_FLOAT, MAX_DOUBLE_DIGIT_LENGTH, "", F_UNSCALED, F_UNSCALED, 2),
-	NUMERIC ("FLOAT", JDBC_REAL, MAX_FLOAT_DIGIT_LENGTH, "", F_UNSCALED, F_UNSCALED, 2),
-	NUMERIC ("DOUBLE PRECISION", JDBC_DOUBLE, MAX_DOUBLE_DIGIT_LENGTH, "", F_UNSCALED, F_UNSCALED, 2),
+	NUMERIC ("DOUBLE PRECISION", JDBC_FLOAT, MAX_DOUBLE_DIGIT_LENGTH, "", UNSCALED, UNSCALED, 2),
+	NUMERIC ("FLOAT", JDBC_REAL, MAX_FLOAT_DIGIT_LENGTH, "", UNSCALED, UNSCALED, 2),
+	NUMERIC ("DOUBLE PRECISION", JDBC_DOUBLE, MAX_DOUBLE_DIGIT_LENGTH, "", UNSCALED, UNSCALED, 2),
 	NUMERIC ("BIGINT", JDBC_BIGINT, MAX_QUAD_LENGTH,"", 0, MAX_QUAD_LENGTH, 10),
-	NUMERIC ("BOOLEAN", JDBC_BOOLEAN, MAX_BOOLEAN_LENGTH, "", F_UNSCALED, F_UNSCALED, F_NOT_NUMERIC),
+	NUMERIC ("BOOLEAN", JDBC_BOOLEAN, MAX_BOOLEAN_LENGTH, "", UNSCALED, UNSCALED, NOT_NUMERIC),
 	DATE("DATE",JDBC_DATE,MAX_DATE_LENGTH,"{d'","'}",1),
 	DATETIME("TIME",JDBC_TIME,MAX_TIME_LENGTH,"{t'","'}",2),
 	DATETIME("TIMESTAMP",JDBC_TIMESTAMP,MAX_TIMESTAMP_LENGTH,"{ts'","'}",3)
 };
-
-static constexpr size_t TYPES_SIZE = sizeof (types) / sizeof (types [0]);
-static constexpr size_t EXPECTED_TYPES_SIZE = 22;
-static_assert( TYPES_SIZE == EXPECTED_TYPES_SIZE, "TypesResultSet.cpp::types array length != 22" );
 
 //////////////////////////////////////////////////////////////////////
 // Construction/Destruction
@@ -144,14 +134,11 @@ static_assert( TYPES_SIZE == EXPECTED_TYPES_SIZE, "TypesResultSet.cpp::types arr
 
 TypesResultSet::TypesResultSet(int dataType, int appOdbcVersion, int bytesPerCharacter, IscConnection* conn) :
 	IscResultSet{nullptr},
-	outputSqlda {conn},
-	status      {conn->GDS->_status},
-	typesOutputHelper{&status, conn->GDS->_master},
-	localBuffer {TYPES_SIZE}
+	outputSqlda {conn}
 {	
 	dataTypes = dataType;
 
-	int endRow = TYPES_SIZE;
+	int endRow = sizeof (types) / sizeof (types [0]);
 
 	if ( appOdbcVersion == 3 ) // SQL_OV_ODBC3
 	{
@@ -200,68 +187,17 @@ TypesResultSet::TypesResultSet(int dataType, int appOdbcVersion, int bytesPerCha
 	types[1].typePrecision = MAX_VARCHAR_LENGTH / bytesPerCharacter;
 	types[2].typePrecision = MAX_BLOB_LENGTH / bytesPerCharacter;
 
-	sqlda = &outputSqlda;
-	sqlda->allocBuffer( statement, typesOutputHelper.getMetadata() );
-
-	auto * typePtr = types;
-
-#define	HLP_SET_STR(FLDNAME,LEN,STR,MAX_SIZE) \
-		typesOutputHelper->FLDNAME.length = typePtr->LEN; \
-		strncpy( typesOutputHelper->FLDNAME.str, typePtr->STR, MAX_SIZE ); \
-		typesOutputHelper->FLDNAME##Null = FB_FALSE;
-
-#define HLP_SET_INT(FLDNAME,VAL) \
-		typesOutputHelper->FLDNAME = typePtr->VAL; \
-		typesOutputHelper->FLDNAME##Null = FB_FALSE;
-
-	for( auto & v : localBuffer )
-	{
-		typesOutputHelper.clear();
-
-		HLP_SET_STR( TYPE_NAME, lenTypeName, typeName, 50 );
-		HLP_SET_INT( DATA_TYPE, typeType );
-		HLP_SET_INT( COLUMN_SIZE, typePrecision );
-		HLP_SET_STR( LITERAL_PREFIX, lenTypePrefix, typePrefix, 6 );
-		HLP_SET_STR( LITERAL_SUFFIX, lenTypeSuffix, typeSuffix, 6 );
-		HLP_SET_STR( CREATE_PARAMS, lenTypeParams, typeParams, 20 );
-		HLP_SET_INT( NULLABLE, typeNullable );
-		HLP_SET_INT( CASE_SENSITIVE, typeCaseSensitive );
-		HLP_SET_INT( SEARCHABLE, typeSearchable );
-		HLP_SET_INT( UNSIGNED_ATTRIBUTE, typeUnsigned );
-		HLP_SET_INT( FIXED_PREC_SCALE, typeMoney );
-		HLP_SET_INT( AUTO_UNIQUE_VALUE, typeAutoIncrement );
-		HLP_SET_STR( LOCAL_TYPE_NAME, lenTypeLocalName, typeLocalName, 50 );
-		HLP_SET_INT( MINIMUM_SCALE, typeMinScale );
-		HLP_SET_INT( MAXIMUM_SCALE, typeMaxScale );
-		HLP_SET_INT( SQL_DATA_TYPE, typeSqlDataType );
-		HLP_SET_INT( SQL_DATETIME_SUB, typeDateTimeSub );
-		HLP_SET_INT( NUM_PREC_RADIX, typeNumPrecRadix );
-		HLP_SET_INT( INTERVAL_PRECISION, typeIntervalPrecision );
-
-		++typePtr;
-
-		auto * buf = (char *)typesOutputHelper.getData();
-		auto buflen = typesOutputHelper.getMetadata()->getMessageLength(&status);
-		v.assign( buf, buf + buflen );
-	}
-	sqlda->buffer.assign( localBuffer[0].begin(), localBuffer[0].end() );	// jump to the first row
-
-#undef	HLP_SET_INT
-#undef	HLP_SET_STR
-
 	recordNumber = 0;
-	numberColumns = sqlda->columnsCount;
+	numberColumns = 19;
 	values.alloc (numberColumns);
 	allocConversions();
 
-	//dark magic here, ptr is out of bounds?..
-	sqldataOffsetPtr = NULL;
-/*
 	indicators = (SQLLEN*)calloc( 1, sizeof(SQLLEN) * numberColumns );
-	((XSQLDA*)*sqlda)->sqld = numberColumns;
+	sqlda = &outputSqlda;
+	sqlda->columnsCount = numberColumns;
+	//((XSQLDA*)*sqlda)->sqld = numberColumns;
 	sqldataOffsetPtr = (uintptr_t)types - sizeof (*types);
 	sqlda->orgsqlvar = new CAttrSqlVar [numberColumns];
-	CAttrSqlVar * orgvar = sqlda->orgsqlvar;
 
 	SET_SQLVAR( 1, "TYPE_NAME"			, SQL_VARYING	,	52  , OFFSET(Types,lenTypeName)				)
 	SET_SQLVAR( 2, "DATA_TYPE"			, SQL_SHORT		,	 5	, OFFSET(Types,typeType)				)
@@ -283,23 +219,21 @@ TypesResultSet::TypesResultSet(int dataType, int appOdbcVersion, int bytesPerCha
 	SET_SQLVAR(18, "NUM_PREC_RADIX"		, SQL_LONG		,	10	, OFFSET(Types,typeNumPrecRadix)		)
 	SET_SQLVAR(19, "INTERVAL_PRECISION"	, SQL_SHORT		,	 5	, OFFSET(Types,typeIntervalPrecision)	)
 
+	CAttrSqlVar * var = sqlda->orgsqlvar;
 	int i = numberColumns;
-	XSQLVAR *var = ((XSQLDA*)*sqlda)->sqlvar;
+	//XSQLVAR *var = ((XSQLDA*)*sqlda)->sqlvar;
 	SQLLEN *ind = indicators;
 
 	while ( i-- )
 	{
-		*orgvar++ = var;
+		//*orgvar++ = var;
 		(var++)->sqlind = (short*)ind++;
 	}
-*/
 }
 
 TypesResultSet::~TypesResultSet()
 {
-	/*
 	free(indicators);
-	*/
 }
 
 bool TypesResultSet::nextFetch()
@@ -315,22 +249,15 @@ bool TypesResultSet::nextFetch()
 			recordNumber = 1;
 			return false;
 		}
+		sqldataOffsetPtr = (uintptr_t)types + (recordNumber - 1) * sizeof (*types);
 	}
 
-	if (++recordNumber > TYPES_SIZE)
+	if (++recordNumber > sizeof (types) / sizeof (types [0]))
 		return false;
 
 	//XSQLVAR *var = ((XSQLDA*)*sqlda)->sqlvar;
-	auto * var = sqlda->orgsqlvar;
-	sqldataOffsetPtr = (size_t)localBuffer[ recordNumber - 1 ].data();
-
-#define SET_INDICATOR_VAL(col,type,isNull) \
-	var[col].assignBuffer( localBuffer[ recordNumber - 1 ].data() ); \
-	if ( isNull && (*(type*)(var[col].sqldata)) == -1 ) *var[col].sqlind = -1;
-
-#define SET_INDICATOR_STR(col) \
-	var[col].assignBuffer( localBuffer[ recordNumber - 1 ].data() ); \
-	if ( *(short*)(var[col].sqldata) == 0 ) *var[col].sqlind = -1;
+	auto *var = sqlda->orgsqlvar;
+	sqldataOffsetPtr += sizeof (*types);
 
 	SET_INDICATOR_STR(0);						// TYPE_NAME
 	SET_INDICATOR_VAL(1,short,false);			// DATA_TYPE
@@ -350,10 +277,7 @@ bool TypesResultSet::nextFetch()
 	SET_INDICATOR_VAL(15,short,false);			// SQL_DATA_TYPE
 	SET_INDICATOR_VAL(16,short,true);			// SQL_DATETIME_SUB
 	SET_INDICATOR_VAL(17,int,true);				// NUM_PREC_RADIX
-	SET_INDICATOR_VAL(18,short,true);			// INTERVAL_PRECISION
-
-#undef SET_INDICATOR_STR
-#undef SET_INDICATOR_VAL
+	SET_INDICATOR_VAL(18,short,true);			// INTERVAL_PRECISION	
 
 	return true;
 }
@@ -373,19 +297,19 @@ bool TypesResultSet::next()
 	++activePosRowInSet;
 
 	//XSQLVAR *var = sqlda->sqlda->sqlvar;
-	//auto * var = sqlda->orgsqlvar;
     Value *value = values.values;
 
+	//for (int n = 0; n < sqlda->sqlda->sqld; ++n, ++var, ++value)
+	//	statement->setValue (value, var);
 	for (int n = 0; n < sqlda->columnsCount; ++n, ++value)
 		statement->setValue (value, n + 1, *sqlda);
-		//TODO: what is "statement" here?.. 
 
 	return true;
 }
 
 int TypesResultSet::findType()
 {	
-	for(int i=0; i < TYPES_SIZE; i++)
+	for(int i=0; i<sizeof (types)/sizeof (types [0]) ; i++)
 		if (types[i].typeType == dataTypes)
 			return i;
